Programmer: Michael Vuong Group: Recreation Outlet (Right Minded) Description: RecRPC is now displayed in the former ID column in the Sales form list view. Accomplishing this also meant changing the addItem methods to take a long id instead of an int id. ID column changed to RPC in the manual item addition.

+++ cs3750Fall2013/RecreationOutlet/Retail/RecreationOutletPOS/RecreationOutletPOS/Combined.cs
@@ -342,7 +342,7 @@ namespace RecreationOutletPOS
         ///     - Converted listview operations into class operations
         ///         (all previous operations were moved to the updateListView function)
         /// </summary>
-        public void addItem(int id, String item, double price, int quantity, double discount, double total)
+        public void addItem(long id, String item, double price, int quantity, double discount, double total)
         {
             tList.addItem(id, item, price, quantity, discount);
             updateListView();
@@ -727,7 +727,7 @@ namespace RecreationOutletPOS
         ///     - Converted listview operations into class operations
         ///         (all previous operations were moved to the updateListView function)
         /// </summary>
-        public void addItem2(int id, String item, double price, int quantity, double discount, double total)
+        public void addItem2(long id, String item, double price, int quantity, double discount, double total)
         {
             tList2.addItem(id, item, price, quantity, discount);
             updateListView2();
